# Improve performance of certain tokenizer states using SWAR