While profiling I noticed that the attribute_value_double_quoted state and data state were very hot in the profile. This patch implements a generic SWAR routine to seek towards an occurrence of any 4 characters in steps of 4/8 bytes at once (depending on the bit width of the CPU).

I benchmarked this on an i7-4790 (32 GiB RAM) using this script: https://gist.github.com/nielsdos/fbbcfb404fdee92e6c814212de81d015 A limited benchmark shows the following:

|                            | baseline            | double_quotes       | double_quotes + state_data |
|----------------------------|---------------------|---------------------|----------------------------|
| StackOverflow homepage     | 329.6 ms ±   3.4 ms | 288.2 ms ±   3.5 ms | 281.0 ms ±   4.6 ms        |
| English Wikipedia homepage | 289.0 ms ±   2.5 ms | 283.5 ms ±   4.5 ms | 275.9 ms ±   4.3 ms        |
| GitHub "For You" page      | 623.2 ms ±   8.0 ms | 570.1 ms ±   4.1 ms | 559.2 ms ±   4.9 ms        |

The baseline was computed on commit e9d35f6384de7bd8c1b79e7111bc3a44f8822967. The double_quotes test was applying SWAR to only the double_quotes state and the last column applies SWAR both for double_quotes and the data state.

As I don't have a dataset, I only tested this on 3 pages.
If you have a benchmark dataset available please let me know so I can evaluate on that.
The limited benchmark shows there might be potential in these optimizations, but doesn't give a proper conclusion yet due to its small size (e.g. are there notable regressions on certain pages?)

I also experimented with using a bitmap instead of SWAR, but that didn't give much of a performance boost.
I also thought about using SSE2, but that's less generically applicable than SWAR and might start to run into a problem of setup overhead.