@lexborisov 2024-04-24
Hi @nielsdos 

I didn't know about this approach, I need to study and test it for some time.
I think I'll have it figured out by Sunday.

Thanks for the patch!