@lexborisov 2024-05-04
@nielsdos 

I had to spend a little more time testing.
In the end, I came to the conclusion that there is no super performance gain here.
Below are my tests, each page parsed `10000` times and outputting the Average and Median value in `sec`.

|                                    | baseline  | double_quotes + state_data |
|------------------------------------|-----------|----------------------------|
| Average habr.com.html              | 0.001763 | 0.001355                    |
| Median habr.com.html               | 0.001750 | 0.001263                    |
| Average amazon.com.html            | 0.000984 | 0.000826                    |
| Median amazon.com.html             | 0.000936 | 0.000776                    |
| Average en.wikipedia.org.html      | 0.000652 | 0.000558                    |
| Median en.wikipedia.org.html       | 0.000738 | 0.000525                    |
| Average ebay.com.html              | 0.000700 | 0.000643                    |
| Median ebay.com.html               | 0.000645 | 0.000644                    |
| Average google.com.html            | 0.000395 | 0.000362                    |
| Median google.com.html             | 0.000363 | 0.000335                    |
| Average youtube.com.html           | 0.000935 | 0.000916                    |
| Median youtube.com.html            | 0.000886 | 0.000855                    |
| Average twitch.com.html            | 0.000185 | 0.000180                    |
| Median twitch.com.html             | 0.000170 | 0.000164                    |
| Average github.com-lexborisov.html | 0.002401 | 0.001761                    |
| Median github.com-lexborisov.html  | 0.002525 | 0.001668                    |

Testing was performed on
|                            | Value                | 
|----------------------------|----------------------|
| Model Name                 | MacBook Pro          |
| Model Identifier           | MacBookPro16,1       |
| Processor Name             | 8-Core Intel Core i9 |
| Processor Speed            | 2,4 GHz              |
| Number of Processors       | 1                    |
| Total Number of Cores      | 8                    |
| L2 Cache (per Core)        | 256 KB               |
| L3 Cache                   | 16 MB                |
| Hyper-Threading Technology | Enabled              |
| Memory                     | 64 GB                |
| System Firmware Version    | 2022.100.22.0.0      |

One thing is certain, the patch does not slow down the code, it speeds it up a bit.
I'm thinking of accepting the patch, and improving SWAR usage a bit in the future.
There are many places in the lexbor project where this approach can be used, for example in CSS.