@nielsdos 2024-05-04
Thanks! In my use case of parsing a lot of pages every improvement counts so I'm glad to see this is accepted :)